feat: 동시성 문제 해결

+++ src/main/java/thegoods/server/item/domain/Item.java
@@ -66,7 +66,7 @@ public class Item extends BaseDateTimeEntity {
     private Long dibsCount;
 
     @Column(nullable = false)
-    private Long salesCount;
+    private Integer salesCount;
 
     @Column
     private Integer tagsCount;
